add permission for book share

+++ app/Repositories/Account/AccountRepository.php
@@ -42,6 +42,7 @@ use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Support\Collection;
+use FireflyIII\Enums\UserRoleEnum;
 
 /**
  * Class AccountRepository.
@@ -55,6 +56,12 @@ class AccountRepository implements AccountRepositoryInterface
      */
     public function destroy(Account $account, ?Account $moveTo): bool
     {
+        // check account manage permission
+        $userGroup     = $this->user->userGroup;
+        $access        = $this->user->hasRoleInGroupOrOwner($userGroup, UserRoleEnum::MANAGE_META) || $this->user->hasRole("owner");
+        if (!$access) {
+            return false;
+        }
         /** @var AccountDestroyService $service */
         $service = app(AccountDestroyService::class);
         $service->destroy($account, $moveTo);
